+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2026-02-04
+This version has some backward **incompatible** changes.
+### Removed
+- Removed virtual `OnDrawGizmosSelected()` from `FlexTargetComponent.cs` since it tends to not be useful at all in real world practice.
+### Changed
+- Changes to `FlexTargetComponent.cs`
+    - `IsTargetValid` no longer has a default implementation. It is now the responsibility of the child class to implement.
+    - Changed the default implementation of `RegisterTargetWhen` to register OnEnable. I think it makes more sense for most situations, since registering OnAwake might add more targets to the repository that constantly need to be filtered out if they are not enabled.
+- Changed builtin targeting data fields to use tooltips from `FlexTargetingTooltips`
+### Added
+- Added static class: `FlexTargetingExtras.Debug` , and static getter: `LastScoredTargetList`. This gives advanced access to the internal list of potential targets and their associated score values that was just populated and sorted by the last called core method. Useful for debugging what score values are being assigned to targets.
+- Added support for the line-of-sight check either using a Raycast or Spherecast. `IFlexData` now has a `LosRaySize` getter. It has a default implementation returning 0, so this change will not break backward compatibility. A value of 0 makes the LOS check use a Raycast. Otherwise, this value will define the sphere radius in a Spherecast.
+    - Also added `_losRaySize` serialized field and getter/setter to all builtin targeting data
+- Added `FlexTargetingTooltips` static class with const strings containing common tooltips for flex data and flex target fields.
+
 ## [0.2.0] - 2026-01-31
 ### Changed
 - Changes to `FlexTargetComponent.cs`